requirements.txt
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.36.1
telethon==1.36.0
python-dotenv==1.0.1
